requirements.txt
Django==2.2.12
django-model-utils==4.0.0
djangorestframework==3.11.0
pytz==2019.1
sqlparse==0.3.0
